stack/airflow/utils/py/bacen/requirements_pip.txt
psycopg2-binary
pandas
python-bcb